src/requirements.txt
asgiref==3.3.4
Django==3.2
djangorestframework==3.12.4
django-filter==2.4.0
django-restql==0.11.2
pytz==2021.1
sqlparse==0.4.1
absl-py==0.11.0
beautifulsoup4==4.9.3
bs4==0.0.1
CacheMan==2.1.0
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cyhunspell==2.0.2
et-xmlfile==1.0.1
filelock==3.0.12
flake8==3.8.4
fsspec==0.8.5
future==0.18.2
gensim==3.8.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
grpcio==1.34.0
idna==2.10
jdcal==1.4.1
joblib==1.0.0
Markdown==3.3.3
mccabe==0.6.1
mypy==0.790
mypy-extensions==0.4.3
nltk==3.5
numpy==1.19.4
oauthlib==3.1.0
openpyxl==3.0.5
packaging==20.8
pandas==1.2.0
protobuf==3.14.0
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pydocstyle==5.1.1
pyflakes==2.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytorch-lightning==1.1.2
PyYAML==5.3.1
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.6
sacremoses==0.0.43
scikit-learn==0.24.0
scipy==1.5.4
six==1.15.0
smart-open==4.1.0
snowballstemmer==2.0.0
soupsieve==2.1
SQLAlchemy==1.4.6
stanza==1.1.1
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
threadpoolctl==2.1.0
tokenizers==0.9.4
torch==1.7.1
torchtext==0.8.1
tqdm==4.55.0
transformers==4.1.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
Unidecode==1.1.2
urllib3==1.26.2
Werkzeug>=2.0